refact: DDD 아키텍처 적용
- 조회가 많은 board 형태의 article의 성능 고도화를 위해 read, write의 분리가 필요하며 usecase와 DIP 구조 적용을 위해 구조 변경
- 상세 구조는 이슈 commit으로 확인

#52
