+++ src/main/java/com/car/sns/config/SecurityConfig.java
@@ -1,24 +1,19 @@
 package com.car.sns.config;
 
-import com.car.sns.dto.UserAccountDto;
-import com.car.sns.dto.security.CarAppPrincipal;
-import com.car.sns.repository.UserAccountRepository;
+import com.car.sns.domain.user.model.UserAccountDto;
+import com.car.sns.security.security.CarAppPrincipal;
+import com.car.sns.domain.user.repository.UserAccountRepository;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.PathResource;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.server.authentication.logout.DelegatingServerLogoutHandler;
-import org.springframework.security.web.server.authentication.logout.SecurityContextServerLogoutHandler;
-import org.springframework.security.web.server.authentication.logout.WebSessionServerLogoutHandler;
 
 @Configuration
 public class SecurityConfig {
