--- src/main/java/com/car/sns/domain/AuditingFields.java
@@ -1,46 +1,0 @@
-package com.car.sns.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import lombok.ToString;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import java.time.LocalDateTime;
-
-@Getter
-@ToString
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public abstract class AuditingFields {
-
-    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-    @CreatedDate
-    @Column(name = "created_at",
-            updatable = false)
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(name = "created_by",
-            length = 100,
-            updatable = false)
-    private String createdBy;
-
-    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-    @LastModifiedDate
-    @Column(name = "modified_at")
-    private LocalDateTime modifiedAt;
-
-    @LastModifiedBy
-    @Column(name = "modified_by",
-            length = 100)
-    private String modifiedBy;
-
-}
-
