+++ src/main/java/com/car/sns/domain/board/repository/querydsl/ArticleRepositoryCustom.java
@@ -1,0 +1,7 @@
+package com.car.sns.domain.board.repository.querydsl;
+
+import java.util.List;
+
+public interface ArticleRepositoryCustom {
+    List<String> findAllDistinctHashtag();
+}
