+++ src/main/java/com/car/sns/domain/board/type/SearchType.java
@@ -1,0 +1,18 @@
+package com.car.sns.domain.board.type;
+
+import lombok.Getter;
+
+public enum SearchType {
+    CONTENT("본문"),
+    HASHTAG("해시태그"),
+    TITLE("제목"),
+    ID("사용자 ID"),
+    NICKNAME("닉네임");
+
+    @Getter
+    private final String description;
+
+    SearchType(String description) {
+        this.description = description;
+    }
+}
