+++ src/main/java/com/car/sns/domain/comment/repository/ArticleCommentRepository.java
@@ -1,0 +1,27 @@
+package com.car.sns.domain.comment.repository;
+
+import com.car.sns.domain.comment.entity.ArticleComment;
+import com.car.sns.domain.QArticleComment;
+import com.querydsl.core.types.dsl.DateTimeExpression;
+import com.querydsl.core.types.dsl.StringExpression;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import org.springframework.data.querydsl.binding.QuerydslBinderCustomizer;
+import org.springframework.data.querydsl.binding.QuerydslBindings;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+@RepositoryRestResource
+public interface ArticleCommentRepository extends
+        JpaRepository<ArticleComment, Long>
+        , QuerydslPredicateExecutor<ArticleComment>
+        , QuerydslBinderCustomizer<QArticleComment> {
+
+    @Override
+    default void customize(QuerydslBindings bindings, QArticleComment root){
+        bindings.excludeUnlistedProperties(true);
+        bindings.including(root.content, root.createdAt, root.createdBy);
+        bindings.bind(root.content).first(StringExpression::containsIgnoreCase);
+        bindings.bind(root.createdAt).first(DateTimeExpression::eq);
+        bindings.bind(root.createdBy).first(StringExpression::containsIgnoreCase);
+    };
+}
