+++ src/main/java/com/car/sns/domain/user/repository/UserAccountRepository.java
@@ -1,0 +1,10 @@
+package com.car.sns.domain.user.repository;
+
+import com.car.sns.domain.user.entity.UserAccount;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
+    Optional<UserAccount> findByUserId(String userId);
+}
