--- src/main/java/com/car/sns/dto/response/ArticleResponse.java
@@ -1,29 +1,0 @@
-package com.car.sns.dto.response;
-
-import com.car.sns.dto.ArticleDto;
-
-import java.time.LocalDateTime;
-
-public record ArticleResponse (
-        String title,
-        String content,
-        String hashtag,
-        LocalDateTime createdAt,
-        String email,
-        String nickname
-) {
-    public static ArticleResponse of(String title, String content, String hashtag, LocalDateTime createdAt, String email, String nickname) {
-        return new ArticleResponse(title, content, hashtag, createdAt, email, nickname);
-    }
-
-    public static ArticleResponse from(ArticleDto dto) {
-        return new ArticleResponse(
-                dto.title(),
-                dto.content(),
-                dto.hashtag(),
-                dto.createdAt(),
-                dto.userAccountDto().email(),
-                dto.userAccountDto().nickname()
-        );
-    }
-}
