--- src/main/java/com/car/sns/dto/security/RoleType.java
@@ -1,14 +1,0 @@
-package com.car.sns.dto.security;
-
-import lombok.Getter;
-
-public enum RoleType {
-    USER("ROLE_USER");
-
-    @Getter
-    private final String name;
-
-    RoleType(String name) {
-        this.name = name;
-    }
-}
