--- src/main/java/com/car/sns/repository/UserAccountRepository.java
@@ -1,10 +1,0 @@
-package com.car.sns.repository;
-
-import com.car.sns.domain.UserAccount;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
-    Optional<UserAccount> findByUserId(String userId);
-}
