--- src/main/java/com/car/sns/repository/querydsl/ArticleRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.car.sns.repository.querydsl;
-
-import java.util.List;
-
-public interface ArticleRepositoryCustom {
-    List<String> findAllDistinctHashtag();
-}
