--- src/main/java/com/car/sns/repository/querydsl/ArticleRepositoryCustomImpl.java
@@ -1,25 +1,0 @@
-package com.car.sns.repository.querydsl;
-
-import com.car.sns.domain.Article;
-import com.car.sns.domain.QArticle;
-import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
-
-import java.util.List;
-
-public class ArticleRepositoryCustomImpl extends QuerydslRepositorySupport implements ArticleRepositoryCustom {
-
-    public ArticleRepositoryCustomImpl() {
-        super(Article.class);
-    }
-
-    @Override
-    public List<String> findAllDistinctHashtag() {
-        QArticle article = QArticle.article;
-
-        return from(article)
-                .distinct()
-                .select(article.hashtag)
-                .where(article.hashtag.isNotNull())
-                .fetch();
-    }
-}
