+++ src/main/java/com/car/sns/security/security/RoleType.java
@@ -1,0 +1,14 @@
+package com.car.sns.security.security;
+
+import lombok.Getter;
+
+public enum RoleType {
+    USER("ROLE_USER");
+
+    @Getter
+    private final String name;
+
+    RoleType(String name) {
+        this.name = name;
+    }
+}
