--- src/main/java/com/car/sns/service/PaginationService.java
@@ -1,27 +1,0 @@
-package com.car.sns.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.IntStream;
-
-@Service
-@RequiredArgsConstructor
-public class PaginationService {
-
-    //bar 길이 : 상태값으로 존재하도록 구현
-    private static final int BAR_LENGTH = 5;
-
-    public List<Integer> getPaginationBarNumbers(int currentPageNumber, int totalPages) {
-        int startNumber = Math.max(currentPageNumber - (BAR_LENGTH / 2), 0);
-        int endNumber = Math.min(startNumber + BAR_LENGTH, totalPages);
-
-        return IntStream.range(startNumber, endNumber).boxed().toList();
-    }
-
-    public int currentBarLength() {
-        return BAR_LENGTH;
-    }
-
-}
