+++ src/test/java/com/car/sns/service/PaginationServiceTest.java
@@ -1,5 +1,6 @@
 package com.car.sns.service;
 
+import com.car.sns.domain.board.service.PaginationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
